tests and fixes.
rending-feed caching
New CacheConfig.java: @EnableCaching + a Caffeine CacheManager with a 60s TTL and a 1,000-entry cap.
New FeedService.java: trending lookup is @Cacheable keyed by {tag, cursor, size}; the personal feed is deliberately not cached.
FeedController.java now delegates to FeedService.
Why: trending ranking is global and identical across users, so it's the one expensive read worth caching. A short TTL absorbs bursts of identical (often anonymous) requests while staying fresh. Personal feeds are per-user and must never be shared, hence the split.
Batched interaction inserts
New InteractionBuffer.java: a bounded in-memory queue drained on a fixed schedule (@Scheduled) via saveAll, with a @PreDestroy flush so buffered rows aren't lost on graceful shutdown. On overflow it persists inline (back-pressure) rather than dropping signals.
InteractionService.java now enqueues into the buffer instead of doing one save per event (still off the request thread via the existing @Async executor).
AsyncConfig.java: added @EnableScheduling for the flush.
application.properties: enabled Hibernate JDBC batching (hibernate.jdbc.batch_size, order_inserts) and added buffer tuning knobs.
Why: interaction tracking is high-volume telemetry. One INSERT per event saturates the connection pool; batching turns N round-trips into a handful of multi-row INSERTs. The trade-off — a crash can lose up to one flush interval of best-effort signals — is acceptable for analytics and is mitigated by the shutdown flush.
Rate limiting on /interactions/track
New RateLimiter.java: per-user token bucket, stored in a size/time-bounded Caffeine cache (no unbounded map growth). No external rate-limit library needed.
New RateLimitExceededException.java; InteractionsController.java checks the limit before processing; GlobalExceptionHandler.java maps it to 429 with a Retry-After header and the standard ErrorResponse body.
openapi.yaml: added a 429 TooManyRequests response to the endpoint so the contract stays aligned with the behavior (consistent with the Pillar-1 emphasis in the original report).
Why: the track endpoint is a trivially spammable firehose into the DB. A per-user bucket bounds abuse; documenting 429 keeps code and spec in sync.
Real integration tests
New InteractionAndFeedIntegrationTests.java: full @SpringBootTest against a real PostgreSQL (Testcontainers) with Flyway building the schema, using MockMvc + the jwt() post-processor — matching the user-service house style. Covers: 202 happy path, 401 unauthenticated, 400 validation (scrollDepth > 100), 429 rate limit, batched-buffer persistence (durationSeconds/scrollDepth round-trip), and trending-feed cache population.
build.gradle: added spring-boot-starter-cache, caffeine, spring-security-test, and Testcontainers (junit-jupiter, postgresql).

+++ backend/recommendation-service/src/main/resources/db/migration/V1__init.sql
@@ -24,7 +24,7 @@ CREATE TABLE interactions (
     post_id          UUID        NOT NULL,
     interaction_type VARCHAR(32) NOT NULL,
     duration_seconds INTEGER     CHECK (duration_seconds >= 0),
-    scroll_depth     SMALLINT    CHECK (scroll_depth BETWEEN 0 AND 100),
+    scroll_depth     INTEGER     CHECK (scroll_depth BETWEEN 0 AND 100),
     metadata         JSONB,
     created_at       TIMESTAMPTZ NOT NULL DEFAULT NOW()
 );
